refactor(frontend): move product to elements folder
